move interface to model/
